SteamPump: change getCapability null side behaviour
Previously if `SteamPump::getCapability` was called with a `null` `side`
then it would invariably return `null`, in conflict with the
`SteamPump::hasCapability` behaviour which defers to the `super`.

This commit brings the two behaviours in sync.

Fixes https://github.com/EmosewaPixel/GregBlock/issues/79.

As a side effect, TOP now shows tank levels for the `SteamPump`. I'm fairly
sure it didn't before.