GTCE version updated to 1.8.0.398
- updated GTCE version to 1.8.0.398
- updated CodeChicken Lib to 3.2.2.353

- RecipeBuilder::chancedOutput
chancedOutput now has an additional parameter, tierChanceBoost, which
has been set to 1000. GTCE source code does not seem to use this value
for anything yet, but usage inspection revealed values of 700-800. It is
likely that a value of 1000 means double per tier, as it used to before
the update.

UnificationEntry::smeltingMultiplier renamed to
UnificationEntry:oreMultiplier

Configurable amperage has been removed from recipes